+++ docs/components/plugins/APISection.test.tsx
@@ -78,8 +78,8 @@ describe('APISection', () => {
     );
 
     expect(screen.getAllByRole('heading', { level: 2 })).toHaveLength(3);
-    expect(screen.getAllByRole('heading', { level: 3 })).toHaveLength(17);
-    expect(screen.getAllByRole('table')).toHaveLength(7);
+    expect(screen.getAllByRole('heading', { level: 3 })).toHaveLength(21);
+    expect(screen.getAllByRole('table')).toHaveLength(9);
 
     expect(screen.queryByText('Classes'));
     expect(screen.queryByText('Asset Properties'));